+++ src/components/project/index.ts
@@ -1,5 +1,7 @@
 export { ProjectHeader } from './ProjectHeader';
 export { ProjectSwitcher } from './ProjectSwitcher';
 export { NewProjectModal } from './NewProjectModal';
+export { EditProjectModal } from './EditProjectModal';
 export { DeleteProjectModal } from './DeleteProjectModal';
 export { ProjectImagesModal } from './ProjectImagesModal';
+export { ProjectOptionsMenu } from './ProjectOptionsMenu';